backend/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
freeze==3.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
six==1.15.0
Werkzeug==1.0.1
psycopg2-binary==2.8.5